update logina i registera
dodato par sitnica za izgled i odradjen deo logike za login i register

+++ src/pages/register/Register.jsx
@@ -1,9 +1,11 @@
-import React from "react";
-import { useForm } from "react-hook-form";
+import React, { useState } from "react";
+import { set, useForm } from "react-hook-form";
 import "./register.scss";
 import { Navigate, useNavigate } from "react-router-dom";
 
 const Register = () => {
+  const [errorMessage, setErrorMessage] = useState("");
+
   const {
     register,
     handleSubmit,
@@ -11,11 +13,21 @@ const Register = () => {
     formState: { errors },
   } = useForm();
   const navigate = useNavigate();
-  const onSubmit = (data) => {
-    localStorage.setItem("test", "test");
-    navigate("/");
+  const onSubmit = async (data) => {
+    // setErrorMessage("");
+    if (watch("password") === watch("confirmpwd")) {
+      localStorage.setItem("test", "test");
+      navigate("/");
+    } else {
+      setErrorMessage("yes");
+      console.log(errorMessage);
+    }
   };
   // console.log(watch("username"));
+
+  const handleLogin = () => {
+    navigate("/login");
+  };
 
   return (
     <section className="registerPage">
@@ -31,27 +43,60 @@ const Register = () => {
           >
             <input
               type="text"
-              {...register("username")}
+              {...register("username", {
+                required: true,
+                minLength: {
+                  value: 4,
+                  message: "Username too short",
+                },
+              })}
               placeholder="username"
             />
             <input
               type="text"
-              {...register("password")}
+              {...register("password", {
+                required: true,
+                minLength: {
+                  value: 4,
+                  message: "Password too short",
+                },
+              })}
               placeholder="password"
             />
             <input
               type="text"
-              {...register("confirmpwd")}
+              {...register("confirmpwd", {
+                required: true,
+              })}
               placeholder="confirm password"
             />
             <input
               type="text"
-              {...register("email", { required: true })}
+              {...register("email", {
+                required: true,
+                pattern: /^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$/,
+              })}
               placeholder="email"
             />
-            {errors.email?.type === "required" && "Email is required"}
+            {errors.username?.type === "required" && (
+              <p>Username is required</p>
+            )}
+            {errors.password?.type === "required" && (
+              <p>Password is required</p>
+            )}
+            {errors.email?.type === "required" && <p>Email is required</p>}
+            {errors.password?.message && <p>{errors.password?.message}</p>}
+            {errors.username?.message && <p>{errors.username?.message}</p>}
+            {watch("password") !== watch("confirmpwd") && (
+                <p>Passwords do not match</p>
+              ) &&
+              errorMessage === "yes"}
 
             <button className="btn">Sign Up</button>
+            <div className="login">
+              Already have an account?
+              <span onClick={handleLogin}> Log in instead!</span>
+            </div>
           </form>
         </div>
       </div>
